# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "One place to know about company events, activities and etc.",
        "version": "1.0.0",
        "title": "Intranet Portal",
        "contact": {
            "email": "sameer_shaik@merilytics.com"
        }
    },
    "host": "localhost:3008",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/resources/getResourcesPolicies": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/searchData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/policiesData"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/getResourcesForms": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/searchData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/policiesData"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/getCategoryDropDown": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Category drop down.",
                "description": "Category drop down.",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resourcesDropDown"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/getFileByID": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/fileOutputs"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/uploadFile": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileUploadInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/deleteFile": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fileDeleteInputs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/resources/getUserApps": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Forms and Policies",
                "description": "Page contains forms and policies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeeId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "employeeId": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "employeeId": 21258
            }
        },
...
        "message": {
            "type": "object",
            "properties": {},
            "example": {
                "message": "Appreciations sent successfully."
            }
        },
...
        "searchData": {
            "type": "object",
            "properties": {},
            "example": {
                "search": null,
                "categoryId": null,
                "pageNo": 1,
                "pageSize": 10
            }
        },
        "policiesData": {
            "type": "object",
            "properties": {},
            "example": {
                "result": [
                    {
                        "rowNumber": "1",
                        "fileId": 11,
                        "fileName": "fee reimbursement",
                        "fileType": "pdf",
                        "resourcesCategoryId": 2,
                        "resourcesCategoryName": "Forms",
                        "createdBy": 21258,
                        "createdByName": "Sameer Shaik",
                        "updatedBy": 21258,
                        "updatedByName": "Sameer Shaik",
                        "createdDate": 1538049733003,
                        "updatedDate": 1538049735717,
                        "totalCount": 3
                    },
                    {
                        "rowNumber": "2",
                        "fileId": 8,
                        "fileName": "cbsdahbc af",
                        "fileType": "pdf",
                        "resourcesCategoryId": 2,
                        "resourcesCategoryName": "Forms",
                        "createdBy": 21280,
                        "createdByName": "Tejeswi Muthireddy",
                        "updatedBy": 21280,
                        "updatedByName": "Tejeswi Muthireddy",
                        "createdDate": 1538045716503,
                        "updatedDate": 1538045718930,
                        "totalCount": 3
                    },
                    {
                        "rowNumber": "3",
                        "fileId": 6,
                        "fileName": "bla bla",
                        "fileType": "pdf",
                        "resourcesCategoryId": 2,
                        "resourcesCategoryName": "Forms",
                        "createdBy": 21258,
                        "createdByName": "Sameer Shaik",
                        "updatedBy": 21258,
                        "updatedByName": "Sameer Shaik",
                        "createdDate": 1537874424977,
                        "updatedDate": 1537874427523,
                        "totalCount": 3
                    }
                ],
                "pages": 1,
                "rowCount": 3
            }
        },
        "resourcesDropDown": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "sortById": 1,
                    "sortingName": "File Name"
                },
                {
                    "sortById": 2,
                    "sortingName": "Uploaded Date"
                }
            ]
        },
        "fileInputs": {
            "type": "object",
            "properties": {},
            "example": {
                "fileId": 4
            }
        },
        "fileOutputs": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "fileId": 4,
                    "resourcesCategoryId": 1,
                    "file": "bla bla",
                    "fileName": "asfjbnfb n fgg",
                    "fileType": "pdf",
                    "createdBy": 21258,
                    "updatedBy": 21258,
                    "createdDate": "2018-09-25T11:19:37.353Z",
                    "updatedDate": "2018-09-25T11:19:41.613Z"
                }
            ]
        },
        "fileUploadInputs": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "resourcesCategoryId": 2,
                    "files": "this is the doc",
                    "fileName": "test doc",
                    "fileType": "pdf",
                    "createdBy": 21258,
                    "updatedBy": 21258
                }
            ]
        },
        "fileDeleteInputs": {
            "type": "object",
            "properties": {},
            "example": {
                "fileId": 13
            }
        },
...
    },
...
}